Move test utils to a common location
