+++ packages/prismic-react/test/Context.test.tsx
@@ -1,25 +1,24 @@
 import React from 'react'
 import { PrismicContext } from '../src'
 import { render } from '@testing-library/react'
-import { testDoc, linkResolver, hrefResolver } from './test-setup'
-
-const Slice: React.FC = () => <div>Hello World</div>
+import { linkResolver, hrefResolver, testDoc, slices } from '../../../test/test-utils'
+import TestSlice from './TestSlice'
 
 test('Context', () => {
   const { container } = render(
     <PrismicContext.Provider value={{
       sliceMap: new Map(Object.entries({
-        slice: Slice,
+        slice: TestSlice,
       })),
       linkResolver,
       hrefResolver,
     }}>
       <PrismicContext.Consumer>
         {(value) => {
-          const SliceComponent = value.sliceMap.get('slice') as React.FC
+          const SliceComponent = value.sliceMap.get('slice') as React.ReactType
           return (
             <div>
-              <SliceComponent />
+              <SliceComponent {...slices[0]}/>
               <div>{value.linkResolver(testDoc)}</div>
               <div>{value.hrefResolver && value.hrefResolver(testDoc)}</div>
             </div>
