+++ packages/prismic-react/test/Slice.test.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 import { act } from '@testing-library/react'
 import { Slice } from '../src/SliceZone/Slice'
-import { renderWithPrismicProvider, slices } from './test-setup'
+import { renderWithPrismicProvider } from './test-setup'
 import TestSlice from './TestSlice'
+import { slices } from '../../../test/test-utils'
+
 
 test('<Slice />', async () => {
   const promise = Promise.resolve()
