+++ packages/prismic-react/test/test-setup.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import { PrismicProvider, urlResolver } from '../src'
+import { PrismicProvider } from '../src'
 import { render, RenderResult } from '@testing-library/react'
 import TestSlice, { NamedTestSlice } from './TestSlice'
 import TestSliceWithState from './TestSliceWithState'
-import { PrismicDoc } from '@stnew/prismic-types'
+import { linkResolver } from '../../../test/test-utils'
 
 type DefaultParams = Parameters<typeof render>;
 type RenderUI = DefaultParams[0];
@@ -20,58 +20,6 @@ const sliceMap = {
   'test_slice_5': ImportedSliceWithState,
   'test_slice_6': NamedImportedSliceWithState,
 }
-
-export const testDoc: PrismicDoc = {
-  type: 'post',
-  uid: 'hello-world',
-  link_type: 'Document',
-  url: '',
-}
-
-export const slices = [
-  {
-    primary: {
-      title: 'Hello World',
-    },
-    items: [],
-    slice_label: null,
-    slice_type: 'test_slice_1',
-  },
-  {
-    primary: {
-      title: 'Hola Mundo',
-    },
-    items: [],
-    slice_label: null,
-    slice_type: 'test_slice_2',
-  },
-  {
-    primary: {
-      title: 'Code-split Component',
-    },
-    items: [],
-    slice_label: null,
-    slice_type: 'test_slice_3',
-  },
-]
-
-export const routes = {
-  'page': {
-    href: '/',
-    page: '/[[...page]]',
-    root: 'homepage',
-  },
-  'blog':{
-    href: '/blog',
-    page: '/blog',
-  },
-  'post': {
-    href: '/blog',
-    page: '/blog/[...uid]',
-  },
-}
-
-export const { linkResolver, hrefResolver } = urlResolver(routes)
 
 export function renderWithPrismicProvider(component: RenderUI): RenderResult {
   return render(
